pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "msg2phone"
version = "0.1.0"
description = "Simple notification helper to push messages to phone."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"requests>=2.0",
	"PyYAML>=6.0",
]

[project.scripts]
msg2phone-cli = "msg2phone.cli:main"

